switched to jsonpickle for serialization

+++ smcpp/commands/estimate.py
@@ -18,7 +18,7 @@ import traceback
 from .. import _smcpp, util, estimation_tools
 from ..model import SMCModel
 from ..population import Population
-from ..inference_service import DumbInferenceService as InferenceService
+from ..inference_service import InferenceService as InferenceService
 from ..optimizer import PopulationOptimizer, TwoPopulationOptimizer
 
 np.set_printoptions(linewidth=120, suppress=True)
@@ -36,7 +36,7 @@ def init_parser(parser):
     hmm.add_argument('--thinning', help="emit full SFS every <k>th site", default=10000, type=int, metavar="k")
     hmm.add_argument('--no-pretrain', help="do not pretrain model", action="store_true", default=False)
     hmm.add_argument('--M', type=int, help="number of hidden states", default=32)
-    hmm.add_argument('--em-iterations', type=float, help="number of EM steps to perform", default=20)
+    hmm.add_argument('--em-iterations', type=int, help="number of EM steps to perform", default=20)
     hmm.add_argument('--regularization-penalty', type=float, help="regularization penalty", default=.01)
     hmm.add_argument('--lbfgs-factor', type=float, help="stopping criterion for optimizer", default=1e9)
     hmm.add_argument('--Nmin', type=float, help="Lower bound on effective population size", default=1000)
